pyproject.toml
[project]
name = "high-quality-blog-finder-site"
version = "0.1.0"
description = "A local-first Chinese high-quality tech blogger discovery site."
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.100",
  "jinja2>=3.1",
  "pyyaml>=6.0",
  "uvicorn>=0.22",
]

[tool.uvicorn]
factory = false
